DS-110: Rename directives
