+++ ngx-fudis/projects/ngx-fudis/src/lib/components/expandable/expandable-directives.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-classes-per-file */
 import { Directive, TemplateRef } from '@angular/core';
 
 /**
@@ -5,5 +6,13 @@ import { Directive, TemplateRef } from '@angular/core';
  */
 @Directive({ selector: '[fudisExpandableContent]' })
 export class ExpandableContentDirective {
+	constructor(public templateRef: TemplateRef<unknown>) {}
+}
+
+/**
+ * A marker directive used to tag the extra buttons rendered in the `ExpandableComponent` header.
+ */
+@Directive({ selector: '[fudisExpandableActions]' })
+export class ExpandableActionsDirective {
 	constructor(public templateRef: TemplateRef<unknown>) {}
 }
