--- ngx-fudis/projects/ngx-fudis/src/lib/components/expandable/expandable-directives/expandable-header-buttons.directive.ts
@@ -1,9 +1,0 @@
-import { Directive, TemplateRef } from '@angular/core';
-
-/**
- * A marker directive used to tag the extra buttons rendered in the `ExpandableComponent` header.
- */
-@Directive({ selector: '[fudisExpandableHeaderButtons]' })
-export class ExpandableHeaderButtonsDirective {
-	constructor(public templateRef: TemplateRef<unknown>) {}
-}
