+++ ngx-fudis/projects/ngx-fudis/src/public-api.ts
@@ -24,8 +24,10 @@ export { DialogService } from './lib/components/dialog/dialog.service';
 export { DropdownComponent } from './lib/components/form/dropdown/dropdown.component';
 // export { ErrorSummaryComponent } from './lib/components/form/error-summary/error-summary.component';
 export { ExpandableComponent } from './lib/components/expandable/expandable.component';
-export { ExpandableContentDirective } from './lib/components/expandable/expandable-directives/expandable-content.directive';
-export { ExpandableHeaderButtonsDirective } from './lib/components/expandable/expandable-directives/expandable-header-buttons.directive';
+export {
+	ExpandableContentDirective,
+	ExpandableActionsDirective,
+} from './lib/components/expandable/expandable-directives';
 export { GridComponent } from './lib/components/grid/grid.component';
 export { GridDirective } from './lib/components/grid/grid.directive';
 export { HeadingComponent } from './lib/components/typography/heading/heading.component';
